--- src/modules/Savannah/components/AudioCard/components/Question/Question.tsx
@@ -1,14 +1,0 @@
-import React, { FC } from 'react';
-import { Word } from 'types';
-import { QuestionWrapper, CurrentWord } from './styled';
-
-type QuestionProps = {
-  word: Word;
-  hasAnswer: boolean;
-};
-
-export const Question: FC<QuestionProps> = ({ word, hasAnswer }) => (
-  <QuestionWrapper>
-    <CurrentWord>{word.word}</CurrentWord>
-  </QuestionWrapper>
-);
