+++ src/modules/Savannah/components/Lives/styled.ts
@@ -1,8 +1,10 @@
-import styled from 'styled-components';
+import styled from 'styled-components/macro';
+import { COLOR_LAYOUT_WHITE } from 'appConstants/colors';
 
 export const LivesWrapper = styled.div`
   position: absolute;
   top: 30px;
   right: 150px;
   display: flex;
+  color: ${COLOR_LAYOUT_WHITE};
 `;
