+++ src/modules/Savannah/components/SavannahCard/styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components/macro';
 
-export const AudioCartStyled = styled.div`
+export const SavannahCartStyled = styled.div`
   width: 100%;
   max-height: 500px;
 `;
