+++ src/modules/Savannah/components/index.ts
@@ -1,3 +1,3 @@
-export { SavannahCard } from './AudioCard';
+export { SavannahCard } from './SavannahCard';
 export { ProgressBar } from './ProgressBar';
 export { Lives } from './Lives';
